refactor GuessNumberConsoleService
